+++ src/main/java/dev/_2lstudios/jelly/player/PluginPlayer.java
@@ -1,8 +1,13 @@
 package dev._2lstudios.jelly.player;
 
+import org.bukkit.Color;
+import org.bukkit.FireworkEffect;
 import org.bukkit.Location;
 import org.bukkit.Sound;
+import org.bukkit.entity.EntityType;
+import org.bukkit.entity.Firework;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.meta.FireworkMeta;
 
 import dev._2lstudios.jelly.utils.ServerUtils;
 
@@ -29,6 +34,21 @@ public class PluginPlayer {
 
     public void sendTitle(final String title, final String subtitle, final int duration) {
         this.sendTitle(title, subtitle, 2, duration * 20, 2);
+    }
+
+    public void spawnFirework(final int amount, final int power, final Color color, final boolean flicker) {
+        final Location loc = this.getBukkitPlayer().getLocation().clone().add(0, 1, 0);
+
+        final Firework firework = (Firework) loc.getWorld().spawnEntity(loc, EntityType.FIREWORK);
+        final FireworkMeta meta = firework.getFireworkMeta();
+        meta.setPower(power);
+        meta.addEffect(FireworkEffect.builder().withColor(color).flicker(flicker).withTrail().build());
+        firework.setFireworkMeta(meta);
+
+        for (int i = 1; i < amount; i++) {
+            final Firework fireworkCopy = (Firework) loc.getWorld().spawnEntity(loc, EntityType.FIREWORK);
+            fireworkCopy.setFireworkMeta(meta);
+        }
     }
 
     @SuppressWarnings("deprecation")