+++ src/main/java/dev/_2lstudios/squidgame/arena/games/G7SquidGame.java
@@ -16,4 +16,9 @@ public class G7SquidGame extends ArenaGameBase {
     public void onTimeUp() {
         this.getArena().killAllPlayers();
     }
+
+    @Override
+    public void onStop() {
+        this.getArena().setPvPAllowed(false);
+    }
 }