+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Blind effect can now be given in game 3
 - Games are now deactivatable by setting their game time to -1
 - Added an option to send the user to the lobby when they join the server.
+- Added fireworks for the arena winner.
 
 ### Fixed